+++ Dyna/Charges/ShatterCharge.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 
 namespace Shockah.Dyna;
 
-public sealed class ShatterCharge : DynaCharge, IRegisterable
+public sealed class ShatterCharge : BaseDynaCharge, IRegisterable
 {
 	private static ISpriteEntry Sprite = null!;
 	private static ISpriteEntry LightsSprite = null!;
@@ -14,6 +14,10 @@ public sealed class ShatterCharge : DynaCharge, IRegisterable
 	{
 		Sprite = ModEntry.Instance.Helper.Content.Sprites.RegisterSprite(ModEntry.Instance.Package.PackageRoot.GetRelativeFile("assets/Charges/Shatter.png"));
 		LightsSprite = ModEntry.Instance.Helper.Content.Sprites.RegisterSprite(ModEntry.Instance.Package.PackageRoot.GetRelativeFile("assets/Charges/ShatterLight.png"));
+	}
+
+	public ShatterCharge() : base($"{ModEntry.Instance.Package.Manifest.UniqueName}::ShatterCharge")
+	{
 	}
 
 	public override Spr GetIcon(State state)