+++ Dyna/IDynaApi.cs
@@ -1,4 +1,5 @@
 ﻿using Nickel;
+using System.Collections.Generic;
 
 namespace Shockah.Dyna;
 
@@ -20,8 +21,35 @@ public interface IDynaApi
 	int GetBlastwaveRange(AAttack attack);
 	AAttack SetBlastwave(AAttack attack, int? damage, int range = 1, bool isStunwave = false);
 
+	IDynaCharge MakeBurstCharge();
+	IDynaCharge MakeConcussionCharge();
+	IDynaCharge MakeDemoCharge();
+	IDynaCharge MakeFluxCharge();
+	IDynaCharge MakeShatterCharge();
+	IDynaCharge MakeSwiftCharge();
+	CardAction MakeFireChargeAction(IDynaCharge charge, int offset = 0, bool targetPlayer = false);
+
+	void DefaultRenderChargeImplementation(IDynaCharge charge, G g, State state, Combat combat, Ship ship, int worldX, Vec position);
+
+	IDynaCharge? GetStickedCharge(State state, Combat combat, Part part);
+	void SetStickedCharge(State state, Combat combat, Part part, IDynaCharge? charge);
+	bool TriggerChargeIfAny(State state, Combat combat, Part part, bool targetPlayer);
+
 	void RegisterHook(IDynaHook hook, double priority);
 	void UnregisterHook(IDynaHook hook);
+}
+
+public interface IDynaCharge
+{
+	string Key();
+	double YOffset { get; set; }
+
+	Spr GetIcon(State state);
+	Spr? GetLightsIcon(State state) => null;
+	void Render(G g, State state, Combat combat, Ship ship, int worldX, Vec position) => ModEntry.Instance.Api.DefaultRenderChargeImplementation(this, g, state, combat, ship, worldX, position);
+
+	IEnumerable<Tooltip> GetTooltips(State state) => [];
+	void OnTrigger(State state, Combat combat, Ship ship, Part part) { }
 }
 
 public interface IDynaHook