+++ Dyna/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Fixed various issues regarding the Ricochet Paddle artifact.
+* Improved Blastwave tooltip text.
 * Added `IDeckEntry DynaDeck` to the API.
 
 ## 1.0.1